Add product support - license generation, info and extraction
Added the -p/--product option.

It receives a parameter of the form:

<PROD>[,VERSION[,FEATURE]*]

Where:
PROD is one of the product names give by the -l/--list option.
VERSION is one of its feature set versions as give by -l PROD
  If not provided - the latest feature set is used
FEATURE is a subset of the VERSION feature set features as listed by
  -l PROD
  If not provided - A full feature set is assumed

All tokens may be case non sensitive unambiguous prefixes and separated
by white-spaces.

If the -p/--product is not provided when creating a license, the license
is unrestricted to any particular product.

Example:

  --product=my_prod,2,feature-1,feature-3

Signed-off-by: Ilan Smith <lunnys@gmail.com>